Save the read only and read write users to MF
