+++ app/models/project_mediaflux.rb
@@ -3,6 +3,9 @@ class ProjectMediaflux
   def self.create!(project:, session_id:, created_by:)
     store_name = Store.default(session_id: session_id).name
 
+    # Make sure the root namespace exists
+    create_root_ns(session_id: session_id)
+
     # Create a namespace for the project
     # The namespace is directly under our root namespace
     project_name = safe_name(project.directory)
@@ -29,6 +32,8 @@ class ProjectMediaflux
       description: project.metadata[:description],
       data_sponsor: project.metadata[:data_sponsor],
       data_manager: project.metadata[:data_manager],
+      data_user_read_only: project.metadata[:data_user_read_only],
+      data_user_read_write: project.metadata[:data_user_read_write],
       departments: project.metadata[:departments]
     }
 
@@ -49,5 +54,19 @@ class ProjectMediaflux
   def self.safe_name(name)
     # only alphanumeric characters
     name.strip.gsub(/[^A-Za-z\d]/, "-")
+  end
+
+  def self.create_root_ns(session_id:)
+    root_namespace = Rails.configuration.mediaflux["api_root_ns"]
+    namespace_request = Mediaflux::Http::NamespaceDescribeRequest.new(path: root_namespace, session_token: session_id)
+    if namespace_request.exists?
+      Rails.logger.info "Root namespace #{root_namespace} already exists"
+    else
+      Rails.logger.info "Created root namespace #{root_namespace}"
+      namespace_request = Mediaflux::Http::NamespaceCreateRequest.new(namespace: root_namespace, description: "TigerData client app root namespace",
+                                                                      store: Store.all(session_id: session_id).first.name,
+                                                                      session_token: session_id)
+      namespace_request.response_body
+    end
   end
 end
